Made a util function for the code running the logid normalization as well as the code finding the most recent log by the command user.
